+++ changelog_entry.yaml
@@ -1,0 +1,4 @@
+- bump: minor
+  changes:
+    added:
+    - Separate tip income and overtime income from the main tax exempt reforms strcuture.
